+++ datastore_server/source/java/ch/systemsx/cisd/openbis/dss/generic/server/DatabaseBasedDataSetPathInfoProvider.java
@@ -53,7 +53,7 @@ public class DatabaseBasedDataSetPathInfoProvider implements IDataSetPathInfoPro
 
         public long size_in_bytes;
         
-        public Long checksum_crc32;
+        public Integer checksum_crc32;
 
         public boolean is_directory;
